Added a simple command and control server.
Currently, this is just a non-block zero MQ socket bound to port
5555 on localhost. It excepts three commands currently; start recording,
stop recording, and add marker to recroding. This is a rough
implementation currently, simple raw strings are sent. Probably will
move to JSON later but I wan't to get more things in place on the client
side before.