Fixed multiple MacOS issues.
Some compile errors in the ControlServer, only showed up in clang.
A bug introduced by updated BASS seems to cause recording to fail
on Mac when 2-channels are specified for a Mic. I think I fixed it
by first trying two channels and if that fails I go for one.